# [Julia] Stablize the results