--- a/run.sh
+++ b/run.sh
@@ -260,9 +260,9 @@ run_julia() {
         julia -e 'using Pkg; Pkg.activate("Related"); Pkg.instantiate()' &&
         if [ $HYPER == 1 ]; then
 
-            capture "Julia" hyperfine -r $runs -w $warmup --show-output "julia --project=Related -e \"using Related; main()\""
+            capture "Julia" hyperfine -r $runs -w $warmup --show-output "julia --startup-file=no --project=Related -e \"using Related; main()\""
         else
-            command ${time} -f '%es %Mk' julia --project=Related -e "using Related; main()"
+            command ${time} -f '%es %Mk' julia --startup-file=no --project=Related -e "using Related; main()"
 
         fi
 
@@ -287,9 +287,9 @@ run_julia_con() {
         cd ./julia_con &&
         julia -e 'using Pkg; Pkg.activate("RelatedCon"); Pkg.instantiate()' &&
         if [ $HYPER == 1 ]; then
-            capture "Julia Concurrent" hyperfine -r $runs -w $warmup --show-output "julia --threads=auto --project=RelatedCon -e \"using RelatedCon; main()\""
+            capture "Julia Concurrent" hyperfine -r $runs -w $warmup --show-output "julia --startup-file=no --threads=auto --project=RelatedCon -e \"using RelatedCon; main()\""
         else
-            command ${time} -f '%es %Mk' julia --threads auto --project=RelatedCon -e "using RelatedCon; main()"
+            command ${time} -f '%es %Mk' julia --startup-file=no --threads=auto --project=RelatedCon -e "using RelatedCon; main()"
         fi
 
     check_output "related_posts_julia_con.json"