@Tortar 2023-11-01
```julia
Interesting...but this doesn't solve actually the problem on my computer...in the 20k posts enabled, this is what I see:

julia> 
bob@bob-Victus-by-HP-Gaming-Laptop-15-fb0xxx:~$ julia # still needs precompilation
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.9.3 (2023-08-24)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> using Related
[ Info: Precompiling Related [892d0954-52c7-4a59-ae8f-9a164aba3227]
Processing time (w/o IO): 4586 milliseconds

julia> main()
Processing time (w/o IO): 267 milliseconds # it went good, consistently these results
13262912

julia> main()
Processing time (w/o IO): 272 milliseconds
^[[A13262912

julia> main()
Processing time (w/o IO): 265 milliseconds
13262912

julia> 
bob@bob-Victus-by-HP-Gaming-Laptop-15-fb0xxx:~$ julia
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.9.3 (2023-08-24)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> using Related

julia> main()
Processing time (w/o IO): 268 milliseconds
13262912

julia> main()
Processing time (w/o IO): 269 milliseconds
13262912

julia> main()
Processing time (w/o IO): 269 milliseconds
13262912

julia> 
bob@bob-Victus-by-HP-Gaming-Laptop-15-fb0xxx:~$ julia # just added a space in the related.jl file to trigger precompilation
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.9.3 (2023-08-24)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> using Related
[ Info: Precompiling Related [892d0954-52c7-4a59-ae8f-9a164aba3227]
Processing time (w/o IO): 4669 milliseconds

julia> main()
Processing time (w/o IO): 361 milliseconds # it went bad
13262912

julia> main()
Processing time (w/o IO): 364 milliseconds
13262912

julia> main()
Processing time (w/o IO): 360 milliseconds
^[[A13262912

julia> main()
Processing time (w/o IO): 360 milliseconds
13262912

julia> 
bob@bob-Victus-by-HP-Gaming-Laptop-15-fb0xxx:~$ julia
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.9.3 (2023-08-24)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> using Related

julia> main()
Processing time (w/o IO): 360 milliseconds # forever bad
13262912

julia> main()
^[[AProcessing time (w/o IO): 361 milliseconds
13262912

julia> main()
Processing time (w/o IO): 358 milliseconds
13262912

julia> 
bob@bob-Victus-by-HP-Gaming-Laptop-15-fb0xxx:~$ julia
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.9.3 (2023-08-24)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> main()
ERROR: UndefVarError: `main` not defined
Stacktrace:
 [1] top-level scope
   @ REPL[1]:1

julia> using Related

julia> main()
Processing time (w/o IO): 361 milliseconds
13262912

julia> main()
Processing time (w/o IO): 367 milliseconds
13262912

julia> 

```