@Tortar 2023-11-01
To me seems like some sort of bug in precompilation, but maybe I'm wrong