@jinyus 2023-11-02
Main:
`Julia | 27.85 ms | 442.00 ms | 3.88 s`

This PR:
```
Julia:

        Benchmark 1: julia --startup-file=no --project=Related -e "using Related; main()"
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 24 milliseconds
          Time (mean ± σ):      1.005 s ±  0.019 s    [User: 0.889 s, System: 0.220 s]
          Range (min … max):    0.982 s …  1.028 s    10 runs
         
Julia:

        Benchmark 1: julia --startup-file=no --project=Related -e "using Related; main()"
        Processing time (w/o IO): 363 milliseconds
        Processing time (w/o IO): 363 milliseconds
        Processing time (w/o IO): 362 milliseconds
          Time (mean ± σ):      1.458 s ±  0.022 s    [User: 1.322 s, System: 0.240 s]
          Range (min … max):    1.443 s …  1.473 s    2 runs
         
Julia:

        Benchmark 1: julia --startup-file=no --project=Related -e "using Related; main()"
        Processing time (w/o IO): 3168 milliseconds
        Processing time (w/o IO): 3164 milliseconds
        Processing time (w/o IO): 3165 milliseconds
          Time (mean ± σ):      4.629 s ±  0.027 s    [User: 4.429 s, System: 0.294 s]
          Range (min … max):    4.610 s …  4.647 s    2 runs
```